MouseSensitivity slider working + tower particle shoot in enemy direction
when you move the mouse sensitivity slider, the value from that slider changes the mouseSensitivity float in the playerMovement script.

For some reason wrote the particle method to track enemies on the y axis only so it was always shooting straight. removed the line of code and it tracked the transform.position correctly